+++ web-admin/src/features/navigation/breadcrumbs/OrganizationAvatar.svelte
